requirements.txt
# Core RAG Dependencies
# NOTE: sentence-transformers is optional - only needed for local mode
# On Vercel, use requirements-vercel.txt instead

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pypandoc==1.12
openpyxl==3.1.2  # For Excel files

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pydantic>=2.6.3,<3.0.0
pydantic-settings==2.1.0

# File handling
aiofiles==23.2.1

# HTTP requests for LLM API and embeddings
requests==2.31.0
huggingface_hub>=0.20.0

# Database and Authentication
sqlalchemy==1.4.46
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
email-validator==2.1.0  # Required for pydantic EmailStr

# Supabase/PostgreSQL support
psycopg2-binary==2.9.9
pgvector==0.2.3

# Optional: Only for local mode (not needed on Vercel)
# chromadb==0.4.22
# sentence-transformers>=2.7.0
# langchain==0.1.0
# langchain-community==0.0.10
